Adjust for postgresql

+++ src/database/timelimitrule.ts
@@ -1,6 +1,6 @@
 /*
  * server component for the TimeLimit App
- * Copyright (C) 2019 - 2020 Jonas Lochmann
+ * Copyright (C) 2019 - 2021 Jonas Lochmann
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as
@@ -61,7 +61,7 @@ export const attributesVersion1: SequelizeAttributes<TimelimitRuleAttributesVers
   categoryId: { ...idWithinFamilyColumn },
   applyToExtraTimeUsage: { ...booleanColumn },
   dayMaskAsBitmask: {
-    type: Sequelize.TINYINT,
+    type: Sequelize.SMALLINT,
     allowNull: false,
     validate: {
       min: 0,
